// apps/main/src/common/infrastructure/database/types/db.d.ts
export type Decimal = ColumnType<string, number | string>;

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;
// ...
export interface WsVillageEstimationDetails {
  additional_facility_service: Generated<number | null>;
  additional_facility_vaccinator_service: Generated<number | null>;
  additional_outreach_service: Generated<number | null>;
  additional_outreach_vaccinator_service: Generated<number | null>;
  available_facillity_service: Generated<number | null>;
  available_outreach_service: Generated<number | null>;
  available_worker: Generated<number | null>;
  created_at: Generated<Date>;
  created_by: number | null;
  deleted_at: Date | null;
  deleted_by: number | null;
  facility_service_percentage: Generated<Decimal | null>;
  gap_health_worker: Generated<number | null>;
  health_worker_ideal_needs: Generated<number | null>;
  id: Generated<number>;
  microplanning_id: number;
  notes: string | null;
  outreach_service_percentage: Generated<Decimal | null>;
  required_monthly_facility_service: Generated<number | null>;
  required_monthly_outreach_service: Generated<number | null>;
  status: Generated<number | null>;
  updated_at: Generated<Date>;
  updated_by: number | null;
  village_id: number;
}

export interface DB {
// ...
  ws_village_estimation_details: WsVillageEstimationDetails;
}